rename local_to_gcs
